+++ package.json
@@ -12,12 +12,13 @@
     "deploy": "vite build && gh-pages -d dist"
   },
   "dependencies": {
+    "@element-plus/icons-vue": "^2.3.1",
+    "@emailjs/browser": "^4.1.0",
     "@fullcalendar/core": "^6.1.11",
     "@fullcalendar/daygrid": "^6.1.11",
     "@fullcalendar/interaction": "^6.1.11",
+    "@fullcalendar/rrule": "^6.1.11",
     "@fullcalendar/vue3": "^6.1.11",
-    "@element-plus/icons-vue": "^2.3.1",
-    "@emailjs/browser": "^4.1.0",
     "@popperjs/core": "^2.11.8",
     "@vee-validate/i18n": "^4.12.5",
     "@vee-validate/rules": "^4.12.5",
